[Fixes #1242] Moves IStartupConfigureServicesFilter and IStartupConfigureContainerFilter interfaces to the internal namespace
