+++ src/Microsoft.AspNetCore.Hosting.Abstractions/Internal/IStartupConfigureContainerFilter.cs
@@ -3,8 +3,12 @@
 
 using System;
 
-namespace Microsoft.AspNetCore.Hosting
+namespace Microsoft.AspNetCore.Hosting.Internal
 {
+    /// <summary>
+    /// This API supports the ASP.NET Core infrastructure and is not intended to be used
+    /// directly from your code. This API may change or be removed in future releases.
+    /// </summary>
     public interface IStartupConfigureContainerFilter<TContainerBuilder>
     {
         Action<TContainerBuilder> ConfigureContainer(Action<TContainerBuilder> container);
