+++ src/Microsoft.AspNetCore.Hosting.Abstractions/Internal/IStartupConfigureServicesFilter.cs
@@ -4,8 +4,12 @@
 using System;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Microsoft.AspNetCore.Hosting
+namespace Microsoft.AspNetCore.Hosting.Internal
 {
+    /// <summary>
+    /// This API supports the ASP.NET Core infrastructure and is not intended to be used
+    /// directly from your code. This API may change or be removed in future releases.
+    /// </summary>
     public interface IStartupConfigureServicesFilter
     {
         Action<IServiceCollection> ConfigureServices(Action<IServiceCollection> next);
